requirements.txt
grpcio
grpcio-tools
tensorflow==2.12 
tf-keras
pymilvus
deepface
opencv-python
redis
